@@ -77,74 +54,57 @@
       "config/**/*.js",
       "hooks/**/*.js",
       "lib/**/*.js",
-      "node_modules/@lando/**/*.cjs",
-      "node_modules/@lando/**/*.js",
       "utils/**/*.js"
     ],
     "assets": [
-      "FATCORE",
       "config.yml",
       "package.json",
       "config/test.yaml",
-      "config/user.yaml",
-      "node_modules/@lando/**/fastcgi_params",
-      "node_modules/@lando/**/*.cnf",
-      "node_modules/@lando/**/*.conf",
-      "node_modules/@lando/**/*.ini",
-      "node_modules/@lando/**/*.inc",
-      "node_modules/@lando/**/*.json",
-      "node_modules/@lando/**/*.php",
-      "node_modules/@lando/**/*.ps1",
-      "node_modules/@lando/**/*.py",
-      "node_modules/@lando/**/*.sh",
-      "node_modules/@lando/**/*.toml",
-      "node_modules/@lando/**/*.tpl",
-      "node_modules/@lando/**/*.tmpl",
-      "node_modules/@lando/**/*.vcl",
-      "node_modules/@lando/**/*.xml",
-      "node_modules/@lando/**/*.yaml",
-      "node_modules/@lando/**/*.yml",
-      "node_modules/figlet/fonts/Small Slant.flf",
-      "node_modules/figlet/fonts/ANSI Shadow.flf"
+      "config/user.yaml"
     ]
   },
   "dependencies": {
     "@lando/argv": "^1.1.2",
-    "@npmcli/arborist": "^6.2.9",
+    "@npmcli/arborist": "^9.0.1",
     "@oclif/core": "^4.2.10",
-    "axios": "0.28.0",
-    "camelcase-keys": "^7",
+    "axios": "^1.8.3",
+    "camelcase-keys": "^9.1.3",
     "copy-dir": "^0.4.0",
-    "dockerode": "^2.4.2",
+    "debug": "^4.4.0",
+    "dockerode": "^4.0.4",
     "fs-extra": "^11.1.1",
+    "get-package-type": "^0.1.0",
     "is-class": "^0.0.9",
-    "is-root": "^2.1.0",
-    "js-yaml": "^4.1.0",
-    "jsonfile": "^2.4.0",
+    "is-interactive": "^2.0.0",
+    "is-root": "^3.0.0",
+    "jsonfile": "^6.1.0",
     "kebabcase-keys": "^1.0.0",
-    "lodash": "^4.17.21",
-    "nanoid": "^3",
+    "lodash-es": "^4.17.21",
+    "nanoid": "^3.3.9",
     "nconf": "^0.12.0",
     "node-cache": "^4.1.1",
-    "pacote": "^15.1.3",
+    "pacote": "^21.0.0",
     "semver": "^7.3.2",
     "slugify": "^1.6.5",
     "string-argv": "0.1.1",
+    "valid-path": "^2.1.0",
     "valid-url": "^1.0.9",
-    "which": "^2.0.2",
-    "yaml": "^2.1.1",
-    "yargonaut": "^1.1.4",
-    "yargs": "^15.4.1"
+    "which": "^5.0.0",
+    "yaml": "^2.7.0"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.16.0",
     "@lando/leia": "^1.0.0-beta.4",
     "@lando/vitepress-theme-default-plus": "^1.1.1",
-    "eslint": "^7.32.0",
-    "eslint-config-google": "^0.9.1",
-    "eslint-plugin-vue": "^8.0.3",
-    "mocha": "^11.1.0",
-    "nyc": "^15.1.0",
+    "@typescript-eslint/eslint-plugin": "^8.26.1",
+    "@typescript-eslint/parser": "^8.26.1",
+    "eslint": "^8.57.1",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-import-resolver-exports": "^1.0.0-beta.5",
+    "eslint-import-resolver-typescript": "^3.8.7",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-vue": "^10.0.0",
+    "prettier": "^3.5.3",
     "vitepress": "^1.6.3"
   }
 }
