@@ -1,29 +1,56 @@
-'use strict';
+import fs from 'node:fs';
+import path from 'node:path';
+import url from 'node:url';
+import { format, inspect } from 'node:util';
 
-// modules
-const chalk = require('chalk');
-const fs = require('fs');
-const gflags = require('../config/flags');
-const os = require('os');
-const path = require('path');
-const yargs = require('yargs');
+import Debug from 'debug';
+import chalk from 'chalk';
+import isObject from 'lodash-es/isPlainObject.js';
+import { Config as OConfig } from '@oclif/core/config';
+import { Errors, Flags } from '@oclif/core';
+import { normalizeArgv } from '@oclif/core/help';
+import { parse } from '@oclif/core/parser';
+// import { ux, Errors } from '@oclif/core';
 
-// OCLIF pieces we need
-const {format, inspect} = require('util');
-const {normalizeArgv} = require('@oclif/core/help');
-const {parse} = require('@oclif/parser');
-const {ux, Errors} = require('@oclif/core');
+import Config from '../lib/config.js';
+import createDebug from '../lib/debug.js';
+import getCacheDir from '../utils/get-cache-dir.js';
+import getObjectSizes from '../utils/get-object-sizes.js';
+import getSize from '../utils/get-size.js';
+import getSystemDataDir from '../utils/get-system-data-dir.js';
+import normalizeManifestPaths from '../utils/normalize-manifest-paths.js';
+import traverseUp from '../utils/traverse-up.js';
+import getDefaultConfig from '../utils/get-default-config.js';
+import runHook from '../utils/run-hook.js';
 
-const OConfig = require('@oclif/core').Config;
+// get __dirname
+const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
 
-// global options
-const globalOptions = gflags.yargs;
+// default consts
+const PRODUCT_ID = path.basename(process.argv[1]) || 'devtool';
+const PRODUCT_ROOT = path.dirname(traverseUp(['package.json'], __dirname).find(fs.existsSync));
+
+// default cli options
+const defaultOptions = {
+  cache: true,
+  cacheDir: getCacheDir(PRODUCT_ID),
+  configTemplates: {
+    global: path.resolve(PRODUCT_ROOT, 'config', 'global.cjs'),
+    user: path.resolve(PRODUCT_ROOT, 'config', 'user.yaml'),
+  },
+  cid: PRODUCT_ID,
+  debug: createDebug(PRODUCT_ID),
+  enableDebugger: false,
+  hooks: {},
+  plugins: [],
+  root: PRODUCT_ROOT,
+};
 
 /*
  * swallows stdout epipe errors
  * this occurs when stdout closes such as when piping to head
  */
-process.stdout.on('error', err => {
+process.stdout.on('error', (err) => {
   if (err && err.code === 'EPIPE') return;
   throw err;
 });
@@ -31,52 +58,101 @@ process.stdout.on('error', err => {
 /*
  * Construct the CLI
  */
-module.exports = class Cli {
-  static id = 'lando';
-  static debug = require('debug')('@lando/cli');
+export default class Cli {
+  static id = PRODUCT_ID;
+  static defaults = defaultOptions;
+  static globalFlags = {
+    'clear': Flags.boolean({ hidden: true }),
+    'config': Flags.string({
+      char: 'c',
+      description: 'use configuration from specified file',
+      env: 'LANDO_CONFIG_FILE',
+      default: undefined,
+      hidden: true,
+      helpGroup: 'GLOBAL',
+    }),
+    'debug': Flags.boolean({ hidden: false }),
+    'no-cache': Flags.boolean({ hidden: false }),
+  };
 
-  #_cache
+  #_cache;
+  #_oclif;
 
   constructor({
     id = Cli.id,
-    cache = true,
-    cacheDir = path.join(os.tmpdir(), Cli.id),
-    cid = Cli.id,
-    debug = Cli.debug,
-    hooks = {},
+    cache = Cli.defaults.cache,
+    cacheDir = Cli.defaults.cacheDir,
+    configTemplates = Cli.defaults.configTemplates,
+    cid = Cli.defaults.cid,
+    debug = Cli.defaults.debug,
+    enableDebugger = Cli.defaults.enableDebugger,
+    hooks = Cli.defaults.hooks,
+    plugins = Cli.defaults.plugins,
+    root = Cli.defaults.root,
   } = {}) {
-    // props
+    // id
+    this.id = id;
+
+    // other propz
     this.cache = cache;
+    this.cacheDir = cacheDir;
+    this.cid = this.id || cid;
+    this.configTemplates = configTemplates;
     this.debug = debug;
     this.hooks = hooks;
-    this.id = id || path.basename(process.argv[1]) || 'lando';
-    this.cid = this.id || cid;
+    this.plugins = plugins;
+    this.root = root;
 
-    // select correct storage backend and setup cache
-    this.StorageBackend = this.cache ? require('../components/file-storage') : require('../components/no-storage');
-    this.#_cache = new this.StorageBackend(({debug: debug.extend('#cache'), dir: cacheDir}));
-    // if cache is disabled and cachedir exist then flush it
-    if (!this.cache && fs.existsSync(cacheDir)) this.StorageBackend.flush(cacheDir, debug.extend('#flush'));
+    // enable debug if needed
+    if (enableDebugger !== false) this.enableDebugger(enableDebugger);
+
+    // computed stuff?
 
     // add the CLIUX module from OCLIF
-    this.ux = ux;
+    // @TODO: this should become some kind of getUX method?
+    // this.ux = ux;
 
     // some debugging about what happened
-    this.debug('instantiated cli with cache %o and hooks %o', this.cache, require('../utils/get-object-sizes')(this.hooks));
+    this.debug('instantiated cli %o with %O', this.id, {
+      cache: this.cache,
+      cacheDir: this.cacheDir,
+      configTempltes: this.configTemplates,
+      hooks: this.hooks,
+      plugins: this.plugins,
+      root: this.root,
+    });
   }
 
-  /**
-   * Returns a parsed array of CLI arguments and options
-   *
-   * @since 3.0.0
-   * @alias lando.cli.argv
-   * @return {Object} Yarg parsed options
-   * @example
-   * const argv = lando.cli.argv();
-   * @todo make this static and then fix all call sites
-   */
-  argv() {
-    return require('yargs').help(false).version(false).argv;
+  async #getStorageBackend(cache = this.cache) {
+    // @IDEA: get this from core plugin manifest?
+    const backends = {
+      'file-storage': () => import('../components/file-storage.js'),
+      'no-storage': () => import('../components/no-storage.js'),
+    };
+
+    const { default: StorageBackend } = await backends[cache === true ? 'file-storage' : 'no-storage']();
+
+    return StorageBackend;
+  }
+
+  async #generateConfigTemplates(templates = [], genOpts = {}) {
+    // filter out stuff so we only end up with stuff we actually need to template
+    templates = templates.filter(([source, dest, options = {}]) => {
+      // if there is no source
+      if (source === undefined || source === null || !fs.existsSync(source)) return false;
+      // if we are not overwriting and the dest already exists
+      if (options.overwrite !== true && fs.existsSync(dest)) return false;
+      // templator
+      return true;
+    });
+
+    if (templates.length > 0) {
+      const { default: Templator } = await import('../lib/templator.js');
+      for (const [source, dest, options] of templates) {
+        const templator = new Templator(source, dest, { debug: this.debug.extend('templator'), ...options });
+        templator.generate(genOpts);
+      }
+    }
   }
 
   async catch(err) {
@@ -84,40 +160,18 @@ module.exports = class Cli {
     if (!err.message) throw err;
     try {
       this.ux.action.stop(chalk.bold.red('!'));
-    } catch {}
+    } catch {
+      // allow failures
+    }
 
     throw err;
   }
 
-  /**
-   * Checks to see if lando is running with sudo. If it is it
-   * will exit the process with a stern warning
-   *
-   * @since 3.0.0
-   * @alias lando.cli.checkPerms
-   * @example
-   * lando.cli.checkPerms()
-   */
-  checkPerms() {
-    const sudoBlock = require('sudo-block');
-    sudoBlock(this.makeArt('sudoRun'));
-  }
-
-  /*
-   * Confirm question
-   */
-  confirm(message = 'Are you sure?') {
-    return {
-      describe: 'Auto answer yes to prompts',
-      alias: ['y'],
-      default: false,
-      boolean: true,
-      interactive: {
-        type: 'confirm',
-        default: false,
-        message: message,
-      },
-    };
+  enableDebugger(namespace) {
+    // normalize namespace to a string
+    namespace = typeof namespace === 'string' ? namespace : `${this.id}*`;
+    // activate
+    Debug.enable(namespace);
   }
 
   error(input, options) {
@@ -130,16 +184,16 @@ module.exports = class Cli {
 
   exitError(input, options = {}, exitCode = 1) {
     // get error
-    Errors.error(input, {...options, exit: false});
+    Errors.error(input, { ...options, exit: false });
     // get code
-    exitCode = (options && options.exit) ? options.exit : exitCode;
+    exitCode = options && options.exit ? options.exit : exitCode;
     // exit
     process.exit(exitCode);
   }
 
   async finally(Error) {
     try {
-      const config = Errors.config;
+      const config = Error.config;
       if (config.errorLogger) await config.errorLogger.flush();
     } catch (error) {
       console.error(error);
@@ -149,66 +203,110 @@ module.exports = class Cli {
   /*
    * Format data
    */
-  formatData(data, {path = '', format = 'default', filter = []} = {}, opts = {}) {
-    return require('./formatters').formatData(data, {path, format, filter}, opts);
-  }
+  // formatData(data, { path = '', format = 'default', filter = [] } = {}, opts = {}) {
+  //   return require('./formatters').formatData(data, { path, format, filter }, opts);
+  // }
 
   /*
    * FormatOptios
    */
-  formatOptions(omit = []) {
-    return require('./formatters').formatOptions(omit);
-  }
+  // formatOptions(omit = []) {
+  //   return require('./formatters').formatOptions(omit);
+  // }
 
-  getHelp(tasks, args, cid = this.cid) {
-    // if we have something cached then just return that
-    if (this.#_cache.has([cid, 'help'])) return this.#_cache.get([cid, 'help']);
+  // getHelp(tasks, args, cid = this.cid) {
+  //   // if we have something cached then just return that
+  //   if (this.#_cache.has([cid, 'help'])) return this.#_cache.get([cid, 'help']);
 
-    // if we get here then we need to do task discovery
-    this.debug('running %o help discovery...', 'cli');
-    // get help
-    const help = tasks.map(task => {
-      // we try catch here because we dont want a busted task to break the whole thing
-      try {
-        return {...require(task.file)(...args), file: task.file};
-      } catch (error) {
-        // @NOTE: what is the best log level for this? warning?
-        this.warn(`Had problems loading task '${task.name}' from ${task.file}!`);
-        this.debug('could not load task %o from %o with error %O', task.name, task.file, error);
-      }
-    })
-    // this makes sure any "caught" tasks dont get added as undefined elements
-    .filter(Boolean)
-    // and then lets sort it alphabetically by command
-    .sort((a, b) => {
-      if (a.command < b.command) return -1;
-      if (a.command > b.command) return 1;
-      return 0;
-    });
+  //   // if we get here then we need to do task discovery
+  //   this.debug('running %o help discovery...', 'cli');
+  //   // get help
+  //   const help = tasks
+  //     .map((task) => {
+  //       // we try catch here because we dont want a busted task to break the whole thing
+  //       try {
+  //         return { ...require(task.file)(...args), file: task.file };
+  //       } catch (error) {
+  //         // @NOTE: what is the best log level for this? warning?
+  //         this.warn(`Had problems loading task '${task.name}' from ${task.file}!`);
+  //         this.debug('could not load task %o from %o with error %O', task.name, task.file, error);
+  //       }
+  //     })
+  //     // this makes sure any "caught" tasks dont get added as undefined elements
+  //     .filter(Boolean)
+  //     // and then lets sort it alphabetically by command
+  //     .sort((a, b) => {
+  //       if (a.command < b.command) return -1;
+  //       if (a.command > b.command) return 1;
+  //       return 0;
+  //     });
 
-    // set and return
-    this.#_cache.set([cid, 'help'], help);
-    return help;
-  }
+  //   // set and return
+  //   this.#_cache.set([cid, 'help'], help);
+  //   return help;
+  // }
 
-  getHooks(ctx) {
-    return ctx.manifest.getUncoded('hooks.cli', {ams: 'aoa'});
-  }
+  // getHooks(ctx) {
+  //   return ctx.manifest.getUncoded('hooks.cli', { ams: 'aoa' });
+  // }
 
-  getTasks(ctx, cid = this.cid) {
-    // if we have something cached then just return that
-    if (this.#_cache.has([cid, 'tasks'])) return this.#_cache.get([cid, 'tasks']);
+  // getTasks(ctx, cid = this.cid) {
+  //   // if we have something cached then just return that
+  //   if (this.#_cache.has([cid, 'tasks'])) return this.#_cache.get([cid, 'tasks']);
 
-    // if we get here then we need to do task discovery
-    this.debug('running %o task discovery...', 'cli');
-    // get the list
-    const tasks = Object.entries(ctx.manifest.get('tasks'))
-      .map(([name, file]) => ({name, file}))
-      .filter(task => fs.existsSync(`${task.file}.js`) || fs.existsSync(task.file));
+  //   // if we get here then we need to do task discovery
+  //   this.debug('running %o task discovery...', 'cli');
+  //   // get the list
+  //   const tasks = Object.entries(ctx.manifest.get('tasks'))
+  //     .map(([name, file]) => ({ name, file }))
+  //     .filter((task) => fs.existsSync(`${task.file}.js`) || fs.existsSync(task.file));
 
-    // set and return
-    this.#_cache.set([cid, 'tasks'], tasks);
-    return tasks;
+  //   // set and return
+  //   this.#_cache.set([cid, 'tasks'], tasks);
+  //   return tasks;
+  // }
+
+  async init() {
+    // @TODO: allow this to only run once eg if #_oclif is set then just return that?
+    // @TODO: do we need anything dynamic from oclif? hooks?
+    // @TODO: do we want to allow plugins?
+
+    // get the oclif config
+
+    const oclif = await OConfig.load(this.root);
+    console.log(oclif);
+    process.exit(1);
+
+    // parse flags and args as best we can
+
+    // augment with additional properties
+    oclif.id = this.id;
+    oclif.cache = this.cache;
+    oclif.configCache = path.join(oclif.cacheDir, `${this.id}-config.json`);
+    oclif.coreDir = this.root;
+    oclif.debuggerEnabled = this.debug.enabled;
+    oclif.env = this.id.toUpperCase();
+    oclif.logger = this.debug;
+    oclif.sysdataDir = getSystemDataDir(this.id);
+
+    // add in some config info
+    oclif.sysConfigFile = path.join(oclif.sysdataDir, 'config.json');
+    oclif.globalConfigFile = path.join(oclif.dataDir, 'config.json');
+    oclif.userConfigFile = path.join(oclif.configDir, 'config.yaml');
+
+    if (isObject(this.configTemplates) && getSize(this.configTemplates) > 0) {
+      await this.#generateConfigTemplates(
+        [
+          [this.configTemplates.system, oclif.sysConfigFile],
+          [this.configTemplates.global, oclif.globalConfigFile, { overwrite: true }],
+          [this.configTemplates.user, oclif.userConfigFile],
+        ],
+        { funcOpts: oclif, readOpts: { fullYamlDoc: true } },
+      );
+    }
+
+    // return
+    return oclif;
   }
 
   log(message = '', ...args) {
@@ -232,8 +330,21 @@ module.exports = class Cli {
    * @example
    * console.log(lando.cli.makeArt('secretToggle', true);
    */
-  makeArt(func, opts) {
-    return require('./art')[func](opts);
+  // makeArt(func, opts) {
+  //   return require('./art')[func](opts);
+  // }
+
+  // @TODO: add support for options.args?
+  async parse(argv = process.argv.slice(2), options = {}) {
+    // NOTE: we are only interested in parsing and normalizing so its ok for validation to fail
+    try {
+      options._parsed = await parse(argv, { strict: false, flags: Cli.globalFlags, ...options });
+    } catch (error) {
+      options._parsed = error.parse.output;
+    }
+
+    // return the same as oclif for compatibility
+    return options._parsed;
   }
 
   /*
@@ -253,124 +364,152 @@ module.exports = class Cli {
    * // Add a task to the yargs CLI
    * yargs.command(lando.tasks.parseToYargs(task));
    */
-  async parseToYargs(task) {
-    const handler = async argv => {
-      // if run is not a function then we need to get it
-      if (task.run instanceof Function === false) task.run = require(task.file)(this).run;
-      // immediately build some arg data set opts and interactive options
-      const data = {options: argv, inquiry: require('./formatters').getInteractive(task.options, argv)};
+  // async parseToYargs(task) {
+  //   const handler = async (argv) => {
+  //     // if run is not a function then we need to get it
+  //     if (task.run instanceof Function === false) task.run = require(task.file)(this).run;
+  //     // immediately build some arg data set opts and interactive options
+  //     const data = { options: argv, inquiry: require('./formatters').getInteractive(task.options, argv) };
 
-      // run our pre command hook
-      // @TODO: add command names and such?
-      await this.runHook('prerun', {id: argv._[0], data, cli: this, task});
+  //     // run our pre command hook
+  //     // @TODO: add command names and such?
+  //     await this.runHook('prerun', { id: argv._[0], data, cli: this, task });
 
-      // queue up an extended debugger
-      const debug = this.debug.extend(`task:${task.command}`);
-      debug('start task %o', task.file);
+  //     // queue up an extended debugger
+  //     const debug = this.debug.extend(`task:${task.command}`);
+  //     debug('start task %o', task.file);
 
-      // run the command here
-      let err;
-      let result;
-      try {
-        debug('start task %o with %o', task.command, data);
-        result = await task.run(data.options, {
-          app: this.app,
-          context: this.context,
-          ctx: this.ctx,
-          debug,
-          [this.id]: this.product,
-          product: this.product,
-        });
-      } catch (error) {
-        err = error;
-        await this.catch(error);
-      } finally {
-        await this.finally(err);
-        debug('done task %o', task.command);
-      }
+  //     // run the command here
+  //     let err;
+  //     let result;
+  //     try {
+  //       debug('start task %o with %o', task.command, data);
+  //       result = await task.run(data.options, {
+  //         app: this.app,
+  //         context: this.context,
+  //         ctx: this.ctx,
+  //         debug,
+  //         [this.id]: this.product,
+  //         product: this.product,
+  //       });
+  //     } catch (error) {
+  //       err = error;
+  //       await this.catch(error);
+  //     } finally {
+  //       await this.finally(err);
+  //       debug('done task %o', task.command);
+  //     }
 
-      // run postrun hook
-      // as per the OCLIF docs this ONLY runs if the command succeeds
-      await this.runHook('postrun', {id: argv._[0], result, cli: this});
+  //     // run postrun hook
+  //     // as per the OCLIF docs this ONLY runs if the command succeeds
+  //     await this.runHook('postrun', { id: argv._[0], result, cli: this });
 
-      // Return result
-      return result;
-    };
+  //     // Return result
+  //     return result;
+  //   };
 
-    // Return our yarg command
-    return {
-      command: task.command,
-      describe: task.describe,
-      builder: require('./formatters').sortOptions(task.options),
-      handler,
-    };
-  }
+  //   // Return our yarg command
+  //   return {
+  //     command: task.command,
+  //     describe: task.describe,
+  //     builder: require('./formatters').sortOptions(task.options),
+  //     handler,
+  //   };
+  // }
 
-  prettify(data, {arraySeparator = ', '} = {}) {
-    return require('../utils/prettify')(data, {arraySeparator});
-  }
+  // prettify(data, { arraySeparator = ', ' } = {}) {
+  //   return require('../utils/prettify')(data, { arraySeparator });
+  // }
 
   /*
    * Run the CLI
    */
   async run(argv = process.argv.slice(2), options = {}) {
     this.debug('running cli with argv %o and options %o', argv, options);
-    // add some color
-    const yargonaut = require('yargonaut');
-    yargonaut.style('green').errorsStyle('red');
 
-    // NOTE: newer versinos of oclif changed "strict" as a parser option so we use the older one, this is probably
-    // ok since we only use it here?
-    const {flags} = await parse(argv, {strict: false, flags: gflags.oclif});
+    // parse argv
+    // @TODO: where does this stuff go?
+    // parses options, flags, command, sets into this?
+    const parsed = await this.parse(argv, options);
+
+    // await this.init?
+    // loads oclif config so we can generate config templates
+    // store oclif on private prop for hell of it?
+    // also merges in any hooks?
+    // @NOTE: these templates should not contain dyanmic data like debug or argv?
+    // @NOTE: this stealth maps in oclif stuff?
+    this.#_oclif = await this.#oclifinit();
+
+    console.log('hi', parsed);
+    process.exit(1);
+
+    // await this.load
+    // this.config = await this.load()?
+    // assembles the config
+
+    // get app
+    // get storage
+    // get commands
+    // get help
+
+    // select correct storage backend and setup cache
+    const StorageBackend = await this.#getStorageBackend();
+    this.#_cache = new StorageBackend({ debug: this.debug.extend('#cache'), dir: this.cacheDir });
+    this.debug('caching %o with %o %o', this.cache, this.#_cache.constructor.name, this.#_cache.dir);
+
+    // if cache is disabled and cachedir exist then flush it
+    if (!this.cache && fs.existsSync(this.cacheDir)) StorageBackend.flush(this.cacheDir, this.debug.extend('#flush'));
 
     // debug if flag config file doesnt exist
     // @NOTE: should this be a proper error?
     // @NOTE: should we try to get an absolute path?
-    if (flags.config && !fs.existsSync(flags.config)) {
-      this.debug('tried to load %o into config but it doesnt exist', flags.config);
+    if (this.flags.config && !fs.existsSync(this.flags.config)) {
+      this.debug('tried to load %o into config but it doesnt exist', this.flags.config);
     }
 
-    // handle legacy and now hidden flags for backwards compatibilities sake
-    if (flags.channel || flags.experimental || flags['secret-toggle']) {
-      this.error('--channel, --experimental and --secret-toggle are no longer valid flags');
-    }
-
-    // get the system (oclif) config
-    const oclif = await OConfig.load(Object.keys(options).length === 0 ? __dirname : options);
     // normalize our argv
-    const [id, ...argvSlice] = normalizeArgv(oclif, argv);
-    // add some extra stuff to the config
-    oclif.argv = argv;
-    oclif.cache = this.cache;
-    oclif.configFile = flags.config;
-    oclif.env = this.id.toUpperCase(),
-    oclif.id = this.id;
-    oclif.logger = this.debug;
+    // @TODO: this does not handle non-command stuff well?
+    // @TODO: revisit?
+    const [id, ...argvSlice] = normalizeArgv(this.#_oclif, this.#_oclif.argv);
+
+    // set command and remaining args
+    this.command = id;
+
+    console.log(this, argvSlice);
+    process.exit(1);
+
+    // generate config templates if needed?
+    // @TODO: should this end up in getConfig?
+
+    // rework oclif conf a bit before we merge
+    process.exit(1);
 
     // debug
-    this.debug('oclif config loaded %O', oclif);
     this.debug('running command %o with args %o', id, argvSlice);
 
-    // start by getting our product config, we do this BEFORE getting the product because the product location
-    // could potentially differ based on the config eg if the user changes config.system.data-dir;
-    // @NOTE: do we want to allow oclif or config to be modifiable via hook this early? or is the minstrapper hook
-    // sufficient?
+    // get early oclif hooks, these are special hooks you can use to mutate the cli config before we start
+    // to look for an app
+    const ohooks = oclif
+      .getPluginsList()
+      .map((plugin) => normalizeManifestPaths({ hooks: plugin.hooks }, plugin.root))
+      .map((plugin) => plugin.hooks)
+      .filter((hooks) => Object.keys(hooks).length > 0);
 
-    const config = require('../config/product')(oclif);
-    const product = config.get('system.product-path');
-    const minstrapper = {config, product};
+    if (ohooks.length > 0) this.hooks = Config.merge(this.hooks, ohooks, ['aoa']);
+    this.debug('found early run oclif hooks %o', getObjectSizes(this.hooks));
 
-    // Concat OCLIF hooks from config if applicable, transform to absolute paths
-    // @NOTE: we do this here instead of passing in during instantiation in bin/lando because Config.load is async
-    // and we want to use that so we can accomodate any "oclif" plugins that add hooks'
-    const ohooks = oclif.getPluginsList()
-      .map(plugin => require('../utils/normalize-manifest-paths')({hooks: plugin.hooks}, plugin.root))
-      .map(plugin => plugin.hooks)
-      .filter(hooks => Object.keys(hooks).length > 0);
-    if (ohooks.length > 0) this.hooks = config.merge(this.hooks, ohooks, ['aoa']);
-    this.debug('found early run oclif hooks %o', require('../utils/get-object-sizes')(this.hooks));
+    // start our config collection and load config sources in decreasing priority
+    // @TODO: move hooks up here so we can use them in getConfig()?
+    // @TODO: optimize this somehow? only generate config if needed? use cache?
+    const config = new Config({
+      cached: oclif.configCache,
+      debug: this.debug.extend('config'),
+      env: this.env,
+      id: this.id,
+      managed: 'global',
+    });
 
-    // minstrap hook
+    // pre-config hook
     //
     // because this event runs BEFORE lando is ready you cannot access it through a lando plugin. instead you need to use
     // an OCLIF hook.
@@ -378,123 +517,149 @@ module.exports = class Cli {
     //
     // FWIW if you are interested in modifying things at this level you should probably just get in touch with us
     // see: https://lando.dev/support
-    await this.runHook('minstrap', {minstrapper});
-    this.debug('going to use %o as product', minstrapper.product);
+    await this.runHook('pre-config', { config });
 
-    // use the minstrapper 2 get da product/lando
-    // we call it "Product" here instead of "Lando" because at this point we want to keep it generic
-    // once we are "downstream" we use Lando/lando as a convention even if the product id is not lando
-    const Product = require(minstrapper.product);
-    // renamespace some debuggerrs
-    Product.debug = this.debug.contract(-1).extend(this.id);
-    minstrapper.config.debug = Product.debug.extend('config');
+    // if we have a CLI provided config source then thats first
+    if (oclif.configFile) config.overrides('userfile', Config.read(oclif.configFile), { encode: false });
+    // then load in product envvars
+    config.env(this.id);
+    // then the user config
+    config.file('user', oclif.userConfigFile);
+    // then the global/managed config
+    config.file('global', oclif.globalConfigFile);
+    // then system configuration if it exits
+    if (fs.existsSync(oclif.sysConfigFile)) config.file('system', oclif.sysConfigFile);
+    // then defaults
+    config.defaults('defaults', getDefaultConfig(this.id));
+    // dump cache
+    config.dump();
+    // this.debug('oclif config loaded %O', oclif);
+    await this.runHook('post-config', { config });
+    process.exit(1);
 
-    // instantiate
-    // @NOTE: we pass in Config, Backend etc because its saves us a good .3s which is a lot in a CLI boot
-    this.product = new Product(minstrapper.config, {StorageBackend: this.StorageBackend});
-    this[this.id] = this.product;
+    // const config = getProductConfig(oclif);
 
-    // Run the product bootstrap
-    try {
-      await this.product.bootstrap(this);
-      this.debug('product %o bootstrap completed successfully', this.product.id);
-    } catch (error) {
-      console.error('Bootstrap failed!'); // eslint-disable-line no-console
-      this.exitError(error);
-    }
+    // console.log(getDefaultConfig(this.id));
+    // console.log(config.get());
+    process.exit(1);
 
-    // determine if we have an app or not
-    // @TODO: findApp is pretty CLI specific? shoudl we have getUtil for this?
-    const appfile = this.product.config.get('core.appfile');
-    const appfiles = [`${appfile}.yaml`, `${appfile}.yml`];
-    const appfilePath = this.product.findApp(appfiles, process.cwd());
+    // this.debug('going to use %o as product', minstrapper.product);
 
-    // if we have an file then lets set some things in the config for downstream purposes
-    if (fs.existsSync(appfilePath)) {
-      // Also get our app object
-      const App = this.product.getComponent('core.app');
-      this.debug('discovered a %o app at %o', App.name, path.dirname(appfilePath));
+    // // use the minstrapper 2 get da product/lando
+    // // we call it "Product" here instead of "Lando" because at this point we want to keep it generic
+    // // once we are "downstream" we use Lando/lando as a convention even if the product id is not lando
+    // const Product = require(minstrapper.product);
+    // // renamespace some debuggerrs
+    // Product.debug = this.debug.contract(-1).extend(this.id);
+    // minstrapper.config.debug = Product.debug.extend('config');
 
-      // override some default static properties
-      App.debug = this.debug.contract(-1);
-      this.app = new App({
-        appfile: appfilePath,
-        config: this.product.config,
-        plugins: this.product.plugins,
-        StorageBackend: this.StorageBackend,
-      });
-    }
+    // // instantiate
+    // // @NOTE: we pass in Config, Backend etc because its saves us a good .3s which is a lot in a CLI boot
+    // this.product = new Product(minstrapper.config, { StorageBackend: this.StorageBackend });
+    // this[this.id] = this.product;
 
-    // determine some context stuff
-    this.context = {app: this.app !== undefined, global: this.app === undefined};
-    this.ctx = this.context.app ? this.app : this.product;
-    this.cid = this.context.app ? this.app.id : this.product.id;
-    this.debug('command is running with context %o and cid %o', this.context, this.cid);
+    // // Run the product bootstrap
+    // try {
+    //   await this.product.bootstrap(this);
+    //   this.debug('product %o bootstrap completed successfully', this.product.id);
+    // } catch (error) {
+    //   console.error('Bootstrap failed!'); // eslint-disable-line no-console
+    //   this.exitError(error);
+    // }
 
-    // merge in additinal hooks from the context
-    config.merge(this.hooks, [this.getHooks(this.ctx)], ['concat']);
-    this.debug('found %o hooks %o', 'cli', require('../utils/get-object-sizes')(this.hooks));
-    // get tasks
-    this.tasks = this.getTasks(this.ctx);
-    // get help https://www.youtube.com/watch?v=CpZakOJlRoY&t=30s
-    this.help = this.getHelp(this.tasks, [this]);
+    // // determine if we have an app or not
+    // // @TODO: findApp is pretty CLI specific? shoudl we have getUtil for this?
+    // const appfile = this.product.config.get('core.appfile');
+    // const appfiles = [`${appfile}.yaml`, `${appfile}.yml`];
+    // const appfilePath = this.product.findApp(appfiles, process.cwd());
 
-    // init hook
-    await this.runHook('init', {id, argv: argvSlice});
+    // // if we have an file then lets set some things in the config for downstream purposes
+    // if (fs.existsSync(appfilePath)) {
+    //   // Also get our app object
+    //   const App = this.product.getComponent('core.app');
+    //   this.debug('discovered a %o app at %o', App.name, path.dirname(appfilePath));
 
-    // Initialize
-    const suffix = this.app ? `(${this.app.name}, v4)` : '(v4)';
-    const cmd = !this.product.config.get('system.packaged') ? '$0' : path.basename(process.execPath) || 'lando';
-    const usage = [`Usage: ${cmd} <command> [args] [options] | ${chalk.magenta(suffix)}`];
+    //   // override some default static properties
+    //   App.debug = this.debug.contract(-1);
+    //   this.app = new App({
+    //     appfile: appfilePath,
+    //     config: this.product.config,
+    //     plugins: this.product.plugins,
+    //     StorageBackend: this.StorageBackend,
+    //   });
+    // }
 
-    // Yargs!
-    yargs.usage(usage.join(' '))
-      .example(`${this.id} start`, 'starts up the app in cwd')
-      .example(`${this.id} rebuild --help`, 'displays help about the rebuild command')
-      .example(`${this.id} destroy -y --debug`, 'runs destroy non-interactively and with debug output')
-      .example(`${this.id} --no-cache`, 'disables and wipes cache')
-      .recommendCommands()
-      .showHelpOnFail(false)
-      .wrap(yargs.terminalWidth() * 0.70)
-      .option('channel', globalOptions.channel)
-      .option('clear', globalOptions.clear)
-      .option('debug', globalOptions.debug)
-      .option('experimental', globalOptions.experimental)
-      .option('no-cache', globalOptions['no-cache'])
-      .help(false)
-      .option('lando', globalOptions.lando)
-      .option('help', globalOptions.help)
-      .option('verbose', globalOptions.verbose)
-      .group('clear', chalk.green('Global Options:'))
-      .group('debug', chalk.green('Global Options:'))
-      .group('help', chalk.green('Global Options:'))
-      .group('no-cache', chalk.green('Global Options:'))
-      .version(false);
+    // // determine some context stuff
+    // this.context = { app: this.app !== undefined, global: this.app === undefined };
+    // this.ctx = this.context.app ? this.app : this.product;
+    // this.cid = this.context.app ? this.app.id : this.product.id;
+    // this.debug('command is running with context %o and cid %o', this.context, this.cid);
 
-    // loop through the tasks and add them to the CLI
-    for (const task of this.help) {
-      if (task.handler) yargs.command(task);
-      else yargs.command(await this.parseToYargs(task));
-    }
+    // // merge in additinal hooks from the context
+    // config.merge(this.hooks, [this.getHooks(this.ctx)], ['concat']);
+    // this.debug('found %o hooks %o', 'cli', require('../utils/get-object-sizes')(this.hooks));
+    // // get tasks
+    // this.tasks = this.getTasks(this.ctx);
+    // // get help https://www.youtube.com/watch?v=CpZakOJlRoY&t=30s
+    // this.help = this.getHelp(this.tasks, [this]);
 
-    // try to get the current tasks
-    const current = this.help.find(task => task.command === id);
+    // // init hook
+    // await this.runHook('init', { id, argv: argvSlice });
 
-    // if we cannot get teh current tasks then show help
-    if (!current) {
-      yargs.showHelp();
-      this.log();
-    }
+    // // Initialize
+    // const suffix = this.app ? `(${this.app.name}, v4)` : '(v4)';
+    // const cmd = !this.product.config.get('system.packaged') ? '$0' : path.basename(process.execPath) || 'lando';
+    // const usage = [`Usage: ${cmd} <command> [args] [options] | ${chalk.magenta(suffix)}`];
 
-    // Show help unless this is a delegation command
-    if ((yargs.argv.help || yargs.argv.lando) && current.delegate !== true) {
-      yargs.showHelp('log');
-      this.log();
-      process.exit(0);
-    }
+    // // Yargs!
+    // yargs
+    //   .usage(usage.join(' '))
+    //   .example(`${this.id} start`, 'starts up the app in cwd')
+    //   .example(`${this.id} rebuild --help`, 'displays help about the rebuild command')
+    //   .example(`${this.id} destroy -y --debug`, 'runs destroy non-interactively and with debug output')
+    //   .example(`${this.id} --no-cache`, 'disables and wipes cache')
+    //   .recommendCommands()
+    //   .showHelpOnFail(false)
+    //   .wrap(yargs.terminalWidth() * 0.7)
+    //   .option('channel', globalOptions.channel)
+    //   .option('clear', globalOptions.clear)
+    //   .option('debug', globalOptions.debug)
+    //   .option('experimental', globalOptions.experimental)
+    //   .option('no-cache', globalOptions['no-cache'])
+    //   .help(false)
+    //   .option('lando', globalOptions.lando)
+    //   .option('help', globalOptions.help)
+    //   .option('verbose', globalOptions.verbose)
+    //   .group('clear', chalk.green('Global Options:'))
+    //   .group('debug', chalk.green('Global Options:'))
+    //   .group('help', chalk.green('Global Options:'))
+    //   .group('no-cache', chalk.green('Global Options:'))
+    //   .version(false);
 
-    // YARGZ MATEY
-    yargs.argv;
+    // // loop through the tasks and add them to the CLI
+    // for (const task of this.help) {
+    //   if (task.handler) yargs.command(task);
+    //   else yargs.command(await this.parseToYargs(task));
+    // }
+
+    // // try to get the current tasks
+    // const current = this.help.find((task) => task.command === id);
+
+    // // if we cannot get teh current tasks then show help
+    // if (!current) {
+    //   yargs.showHelp();
+    //   this.log();
+    // }
+
+    // // Show help unless this is a delegation command
+    // if ((yargs.argv.help || yargs.argv.lando) && current.delegate !== true) {
+    //   yargs.showHelp('log');
+    //   this.log();
+    //   process.exit(0);
+    // }
+
+    // // YARGZ MATEY
+    // yargs.argv;
   }
 
   reinit(cid = this.cid) {
@@ -503,11 +668,11 @@ module.exports = class Cli {
   }
 
   async runHook(event, data) {
-    return require('../utils/run-hook')(event, data, this.hooks, {cli: this}, this.debug, this.exitError);
+    return runHook(event, data, this.hooks, { cli: this }, this.debug, this.exitError);
   }
 
   warn(input) {
     Errors.warn(input);
     return input;
   }
-};
+}
