@@ -1,24 +1,28 @@
-'use strict';
+import fs from 'node:fs';
+import path from 'node:path';
 
-// @TODO: remove Config as dep for utils?
-// * get-plugins
-// @TODO: add COnfig to this or new()? or both?
+import get from 'lodash-es/get.js';
+import has from 'lodash-es/has.js';
+import nconf from 'nconf';
+import set from 'lodash-es/set.js';
+import unset from 'lodash-es/unset.js';
 
-// modz
-const fs = require('fs');
-const get = require('lodash/get');
-const has = require('lodash/has');
-const nconf = require('nconf');
-const set = require('lodash/set');
-const unset = require('lodash/unset');
+import { parse as parseYAML, stringify as stringifyYAML } from 'yaml';
 
-const {basename, dirname, extname} = require('path');
+import createDebug from '../lib/debug.js';
+import encode from '../utils/encode.js';
+import encodeKeys from '../utils/encode-keys.js';
+import decodeKeys from '../utils/decode-keys.js';
+import getObjectKeys from '../utils/get-object-keys.js';
+import merge from '../utils/merge.js';
+import read from '../utils/read-file.js';
+import write from '../utils/write-file.js';
 
 // custom yml/yaml formatter
-['yaml', 'yml'].forEach(format => {
+['yaml', 'yml'].forEach((format) => {
   nconf.formats[format] = {
-    parse: (obj, options) => require('js-yaml').load(obj, options),
-    stringify: (obj, options) => require('js-yaml').dump(obj, options),
+    parse: (data, options) => parseYAML(data, options),
+    stringify: (obj, options) => stringifyYAML(obj, options),
   };
 });
 
@@ -27,23 +31,23 @@ const {basename, dirname, extname} = require('path');
  */
 class Config extends nconf.Provider {
   static id = 'config';
-  static debug = require('./debug')('@lando/core:config');
+  static debug = createDebug('devtool:config');
 
-  static keys(data, {prefix = '', expandArrays = true} = {}) {
-    return require('../utils/get-object-keys')(data, {prefix, expandArrays});
+  static keys(data, { prefix = '', expandArrays = true } = {}) {
+    return getObjectKeys(data, { prefix, expandArrays });
   }
 
   // @TODO: error handling?
   static merge(object, sources, ams = ['merge:id', 'replace']) {
-    return require('../utils/merge')(object, sources, ams);
+    return merge(object, sources, ams);
   }
 
   static read(file, options) {
-    return require('../utils/read-file')(file, options);
+    return read(file, options);
   }
 
   static write(file, data, options) {
-    return require('../utils/write-file')(file, data, options);
+    return write(file, data, options);
   }
 
   constructor({
@@ -52,7 +56,7 @@ class Config extends nconf.Provider {
     debug = Config.debug,
     decode = true,
     encode = true,
-    id = Config.id || basename(process.argv[1]),
+    id = Config.id ?? path.basename(process.argv[1]),
     managed = 'managed',
     stores = {},
     cached,
@@ -70,7 +74,7 @@ class Config extends nconf.Provider {
     this.managed = managed;
 
     // start with a memcache
-    this.add('#mem', {type: 'memory', readOnly: false});
+    this.add('#mem', { type: 'memory', readOnly: false });
     // populated with data if applicable
     if (data && data !== null && typeof data === 'object') this.stores['#mem'].store = data;
 
@@ -80,19 +84,19 @@ class Config extends nconf.Provider {
     }
 
     // log
-    this.debug('initialized config %o with options %o', this.id, {cached, decode, encode, managed});
+    this.debug('initialized config %o with options %o', this.id, { cached, decode, encode, managed });
   }
 
   add(name, options) {
     // log
-    const dopts = {encode: get(options, 'encode', this.encode), source: options.file};
+    const dopts = { encode: get(options, 'encode', this.encode), source: options.file };
     this.debug('adding %o store %o to %o config %o', options.type, name, this.id, dopts);
     // kick upstream
-    super.add(name, {logicalSeparator: '.', parseValues: true, ...options});
+    super.add(name, { logicalSeparator: '.', parseValues: true, ...options });
 
     // encode the new store if applicable
     if (get(options, 'encode', this.encode) && get(this, `stores.${name}.store`)) {
-      this.stores[name].store = require('../utils/encode-keys')(this.stores[name].store);
+      this.stores[name].store = encodeKeys(this.stores[name].store);
     }
 
     // ensure store as at least an empty object
@@ -102,11 +106,11 @@ class Config extends nconf.Provider {
   }
 
   defaults(name, data, options) {
-    return this.add(name, {...options, type: 'literal', store: data});
+    return this.add(name, { ...options, type: 'literal', store: data });
   }
 
   dump() {
-    fs.mkdirSync(dirname(this.cached), {recursive: true});
+    fs.mkdirSync(path.dirname(this.cached), { recursive: true });
     Config.write(this.cached, this);
     this.debug('dumped compiled and cached config file to %o', this.cached);
   }
@@ -118,7 +122,7 @@ class Config extends nconf.Provider {
       logicalSeparator: '.',
       lowerCase: true,
       parseValues: true,
-      transform: obj => {
+      transform: (obj) => {
         if (obj.key.startsWith(rootKey.toLowerCase())) {
           obj.key = obj.key.replace(rootKey.toLowerCase(), '');
           obj.key = obj.key.replace(new RegExp(`${separator}${separator}`, 'g'), '-');
@@ -133,11 +137,11 @@ class Config extends nconf.Provider {
 
   file(name, options) {
     // if options is a string and path that exists then reset it as file
-    if (typeof options === 'string' && fs.existsSync(options)) options = {file: options};
+    if (typeof options === 'string' && fs.existsSync(options)) options = { file: options };
 
     // try to get the format
     // @TODO: throw error if format is undefined?
-    const extension = extname(options.file).substring(1);
+    const extension = path.extname(options.file).substring(1);
     const format = nconf.formats[extension];
 
     // if a js file then we can only load as a literal,
@@ -145,12 +149,12 @@ class Config extends nconf.Provider {
     // @NOTE: this can also load JS functions and it injects THIS into them
     // @NOTE: we also do this instead of a formatter because that would require use of eval
     if (extension === 'js') {
-      const store = (typeof Config.read(options.file) === 'function') ? Config.read(options.file)(this) : Config.read(options.file);
-      return this.add(name, {...options, type: 'literal', store});
+      const store = typeof Config.read(options.file) === 'function' ? Config.read(options.file)(this) : Config.read(options.file);
+      return this.add(name, { ...options, type: 'literal', store });
     }
 
     // otherwise just pass through
-    return this.add(name, {...options, format, type: 'file'});
+    return this.add(name, { ...options, format, type: 'file' });
   }
 
   // replace nconf get wrapper with this one
@@ -164,50 +168,52 @@ class Config extends nconf.Provider {
       path = parts.join('.');
       data = this.stores[store].store;
 
-    // otherwise get the merged data, this is where the magix happens
+      // otherwise get the merged data, this is where the magix happens
     } else {
       data = Config.merge(
         {},
-        Object.keys(this.stores).reverse().map(store => this.stores[store].store),
+        Object.keys(this.stores)
+          .reverse()
+          .map((store) => this.stores[store].store),
         get(options, 'ams', this.ams),
       );
     }
 
     // if we are encoded then make sure the path is encoded
-    if (get(options, 'encode', this.encode)) path = require('../utils/encode')(path);
+    if (get(options, 'encode', this.encode)) path = encode(path);
 
     // if we have a path then select from data
     if (path) data = get(data, path);
 
     // return decoded or not
     this.debug('getting %o from %o config', path || 'everything', store ? `${this.id}:${store}` : this.id);
-    return get(options, 'decode', this.decode) ? require('../utils/decode-keys')(data) : data;
+    return get(options, 'decode', this.decode) ? decodeKeys(data) : data;
   }
 
   // like get but with no decoding/encoding
   getRaw(path, options) {
-    return this.get(path, {...options, encode: false, decode: false});
+    return this.get(path, { ...options, encode: false, decode: false });
   }
 
   // return uncoded
   getUncoded(path, options) {
-    return this.get(path, {...options, encode: true, decode: false});
+    return this.get(path, { ...options, encode: true, decode: false });
   }
 
-  keys(data = this.getUncoded(), {prefix = '', expandArrays = true} = {}) {
-    return require('../utils/get-object-keys')(data, {prefix, expandArrays});
+  keys(data = this.getUncoded(), { prefix = '', expandArrays = true } = {}) {
+    return getObjectKeys(data, { prefix, expandArrays });
   }
 
   merge(object, sources, ams = this.ams) {
-    return require('../utils/merge')(object, sources, ams);
+    return merge(object, sources, ams);
   }
 
   newConfig(options) {
-    return new Config({debug: this.debug, ...options});
+    return new Config({ debug: this.debug, ...options });
   }
 
   overrides(name, data, options) {
-    return this.add(name, {...options, type: 'literal', store: data});
+    return this.add(name, { ...options, type: 'literal', store: data });
   }
 
   // removes a property and saves the store if applicable
@@ -216,25 +222,26 @@ class Config extends nconf.Provider {
     if (!has(this.stores, `${store}.file`)) {
       unset(this.stores[store].store, path);
 
-    // otherwise we are talking about files
+      // otherwise we are talking about files
     } else {
       // get the store destination
       const dest = get(this.stores, `${store}.file`);
 
       // if this is a yaml file then lets try to reconcile comments and data
-      if (['.yaml', '.yml'].includes(extname(dest))) {
+      if (['.yaml', '.yml'].includes(path.extname(dest))) {
         // load the yaml doc
-        const yamlDoc = require('yaml').parseDocument(fs.readFileSync(dest, 'utf8'));
+        const yamlDoc = Config.read(dest, { fullYamlDoc: true });
+
         // remove the prop
         yamlDoc.deleteIn(path.split('.'));
 
         // write the result
         Config.write(dest, yamlDoc);
 
-      // otherwise to the usual removal
+        // otherwise to the usual removal
       } else {
         unset(this.stores[store].store, path);
-        Config.write(dest, (this.stores[store].store));
+        Config.write(dest, this.stores[store].store);
       }
     }
 
@@ -268,12 +275,12 @@ class Config extends nconf.Provider {
     if (!dest) throw new Error('something');
 
     // figure out the encoding
-    if (get(options, 'encode', this.encode)) data = require('../utils/encode-keys')(data);
+    if (get(options, 'encode', this.encode)) encodeKeys(data);
 
     // if this is a yaml file then lets try to reconcile comments and data
-    if (['.yaml', '.yml'].includes(extname(dest))) {
+    if (['.yaml', '.yml'].includes(path.extname(dest))) {
       // load the yaml doc
-      const yamlDoc = require('yaml').parseDocument(fs.readFileSync(dest, 'utf8'));
+      const yamlDoc = Config.read(dest, { fullYamlDoc: true });
       // go through the data and set it into the doc
       for (const path of Config.keys(data)) {
         yamlDoc.setIn(path.split('.'), get(data, path));
@@ -282,7 +289,7 @@ class Config extends nconf.Provider {
       // write the result
       Config.write(dest, yamlDoc);
 
-    // otherwise to the usual saving
+      // otherwise to the usual saving
     } else {
       this.stores[store].store = Config.merge({}, [this.stores[store].store, data], get(options, 'ams', this.ams));
       Config.write(dest, this.stores[store].store);
@@ -298,7 +305,7 @@ class Config extends nconf.Provider {
 
   set(path, data, options) {
     // encode the path if needed
-    if (get(options, 'encode', this.encode)) path = require('../utils/encode')(path);
+    if (get(options, 'encode', this.encode)) path = encode(path);
     // then just bounce upstream
     super.set(path, data);
   }
@@ -309,4 +316,4 @@ class Config extends nconf.Provider {
   }
 }
 
-module.exports = Config;
+export default Config;
