@@ -1,13 +1,24 @@
-'use strict';
+import encodeKeys from './encode-keys.js';
+import kebabCase from 'lodash-es/kebabCase.js';
 
 // @TODO: throw error for nully values?
-module.exports = data => {
+export default (data) => {
   // if we have a nully value then just return
   if (data === null || data === undefined) return data;
   // if string then return
-  if (typeof data === 'string') return data.split('.').map(part => require('lodash/kebabCase')(part)).join('.');
+  if (typeof data === 'string')
+    return data
+      .split('.')
+      .map((part) => kebabCase(part))
+      .join('.');
   // if array then map and return
-  if (Array.isArray(data)) return data.map(prop => prop.split('.').map(part => require('lodash/kebabCase')(part)).join('.'));
+  if (Array.isArray(data))
+    return data.map((prop) =>
+      prop
+        .split('.')
+        .map((part) => kebabCase(part))
+        .join('.'),
+    );
   // else assume object and return, ignore keys in plugin format
-  return require('./encode-keys')(data);
+  return encodeKeys(data);
 };
