@@ -1,38 +1,41 @@
-'use strict';
+import os from 'node:os';
+import path from 'node:path';
 
-const os = require('os');
-const path = require('path');
+import getEnvironment from './get-environment.js';
+import getPlatform from './get-platform.js';
+
+const env = getEnvironment();
+const platform = getPlatform();
 
 const getOClifHome = () => {
-  switch (process.platform) {
+  switch (platform) {
     case 'darwin':
     case 'linux':
-      return process.env.HOME || os.homedir() || os.tmpdir();
+    case 'wsl':
+      return env.HOME ?? os.homedir() ?? os.tmpdir();
     case 'win32':
-      return process.env.HOME
-        || (process.env.HOMEDRIVE && process.env.HOMEPATH && path.join(process.env.HOMEDRIVE, process.env.HOMEPATH))
-        || process.env.USERPROFILE
-        || os.homedir()
-        || os.tmpdir();
+      return (
+        env.HOME ||
+        (env.HOMEDRIVE && env.HOMEPATH && path.join(env.HOMEDRIVE, env.HOMEPATH)) ||
+        env.USERPROFILE ||
+        os.homedir() ||
+        os.tmpdir()
+      );
   }
 };
 
 /*
  * Get oclif base dir based on platform
  */
-const getOClifBase= product => {
-  const base = process.env['XDG_CACHE_HOME']
-    || (process.platform === 'win32' && process.env.LOCALAPPDATA)
-    || path.join(getOClifHome(), '.cache');
+const getOClifBase = (product) => {
+  const base = env['XDG_CACHE_HOME'] || (platform === 'win32' && env.LOCALAPPDATA) || path.join(getOClifHome(), '.cache');
   return path.join(base, product);
 };
 
-const macosCacheDir = product => {
-  return process.platform === 'darwin' ? path.join(getOClifHome(), 'Library', 'Caches', product) : undefined;
+const macosCacheDir = (product) => {
+  return platform === 'darwin' ? path.join(getOClifHome(), 'Library', 'Caches', product) : undefined;
 };
 
-module.exports = (product = 'hyperdrive') => {
-  return process.env[`${product.toUpperCase()}_CACHE_DIR`]
-    || macosCacheDir(product)
-    || getOClifBase(product);
+export default (product = 'devtool') => {
+  return env[`${product.toUpperCase()}_CACHE_DIR`] ?? macosCacheDir(product) ?? getOClifBase(product);
 };
