@@ -15,27 +15,29 @@ const BOOTSTRAP_LEVELS = {
 };
 
 // Default version information
-const DEFAULT_VERSIONS = {networking: 1};
+const DEFAULT_VERSIONS = { networking: 1 };
 
 // Helper to get init config
-const getInitConfig = dirs => _(dirs)
-  .filter(dir => fs.existsSync(dir))
-  .flatMap(dir => glob.sync(path.join(dir, '*', 'init.js')))
-  .map(file => require(file))
-  .value();
+const getInitConfig = (dirs) =>
+  _(dirs)
+    .filter((dir) => fs.existsSync(dir))
+    .flatMap((dir) => glob.sync(path.join(dir, '*', 'init.js')))
+    .map((file) => require(file))
+    .value();
 
 // Helper to get init source config
-const getInitSourceConfig = dirs => _(dirs)
-  .filter(dir => fs.existsSync(dir))
-  .flatMap(dir => glob.sync(path.join(dir, '*.js')))
-  .map(file => require(file))
-  .flatMap(source => source.sources)
-  .value();
+const getInitSourceConfig = (dirs) =>
+  _(dirs)
+    .filter((dir) => fs.existsSync(dir))
+    .flatMap((dir) => glob.sync(path.join(dir, '*.js')))
+    .map((file) => require(file))
+    .flatMap((source) => source.sources)
+    .value();
 
 /*
  * Helper to bootstrap plugins
  */
-const bootstrapConfig = lando => {
+const bootstrapConfig = (lando) => {
   const Plugins = require('./plugins');
   lando.plugins = new Plugins(lando.log);
   lando.versions = _.merge({}, DEFAULT_VERSIONS, lando.cache.get('versions'));
@@ -44,61 +46,73 @@ const bootstrapConfig = lando => {
   // Load in experimental features
   if (lando.config.experimental) {
     const experimentalPluginPath = path.join(__dirname, '..', 'experimental');
-    lando.config.pluginDirs.push({path: experimentalPluginPath, subdir: '.'});
+    lando.config.pluginDirs.push({ path: experimentalPluginPath, subdir: '.' });
   }
   // Find the plugins
-  return lando.plugins.find(lando.config.pluginDirs, lando.config)
-  // Init the plugins
-  .map(plugin => lando.plugins.load(plugin, plugin.path, lando))
-  .map(plugin => {
-    // Merge in config
-    if (_.has(plugin, 'data.config')) lando.config = _.merge(plugin.data.config, lando.config);
-    // Add plugins to config
-    // @NOTE: we remove plugin.data here because circular ref error and because presumably that
-    // data is now expessed directly in the lando object somewhere
-    lando.config.plugins.push(_.omit(plugin, 'data'));
-  });
+  return (
+    lando.plugins
+      .find(lando.config.pluginDirs, lando.config)
+      // Init the plugins
+      .map((plugin) => lando.plugins.load(plugin, plugin.path, lando))
+      .map((plugin) => {
+        // Merge in config
+        if (_.has(plugin, 'data.config')) lando.config = _.merge(plugin.data.config, lando.config);
+        // Add plugins to config
+        // @NOTE: we remove plugin.data here because circular ref error and because presumably that
+        // data is now expessed directly in the lando object somewhere
+        lando.config.plugins.push(_.omit(plugin, 'data'));
+      })
+  );
 };
 
 /*
  * Helper to bootstrap tasks
  */
-const bootstrapTasks = lando => {
+const bootstrapTasks = (lando) => {
   // Load in config from init and source plugins
   const inits = getInitConfig(_.map(lando.config.plugins, 'recipes'));
   const sources = getInitSourceConfig(_.map(lando.config.plugins, 'sources'));
-  const initSources = _(inits).filter(init => _.has(init, 'sources')).flatMap(init => init.sources).value();
+  const initSources = _(inits)
+    .filter((init) => _.has(init, 'sources'))
+    .flatMap((init) => init.sources)
+    .value();
   // Sort into useful things and add to config
   lando.config.sources = _.sortBy(sources.concat(initSources), 'label');
-  lando.config.recipes = _.sortBy(_.map(inits, init => init.name), 'name');
+  lando.config.recipes = _.sortBy(
+    _.map(inits, (init) => init.name),
+    'name',
+  );
   lando.config.inits = inits;
 
   // Load in all our tasks
-  return lando.Promise.resolve(lando.config.plugins)
-    // Make sure the tasks dir exists
-    .filter(plugin => fs.existsSync(plugin.tasks))
-    // Get a list off full js files that exist in that dir
-    .map(plugin => _(fs.readdirSync(plugin.tasks))
-      .map(file => path.join(plugin.tasks, file))
-      .filter(path => _.endsWith(path, '.js'))
-      .value(),
-    )
-    // Loadem and loggem
-    .then(tasks => _.flatten(tasks))
-    .each(task => {
-      lando.tasks.push(require(task)(lando));
-      lando.log.debug('autoloaded task %s', path.basename(task, '.js'));
-    })
-    // Reset the task cache
-    .then(() => {
-      lando.cache.set('_.tasks.cache', JSON.stringify(lando.tasks), {persist: true});
-    });
+  return (
+    lando.Promise.resolve(lando.config.plugins)
+      // Make sure the tasks dir exists
+      .filter((plugin) => fs.existsSync(plugin.tasks))
+      // Get a list off full js files that exist in that dir
+      .map((plugin) =>
+        _(fs.readdirSync(plugin.tasks))
+          .map((file) => path.join(plugin.tasks, file))
+          .filter((path) => _.endsWith(path, '.js'))
+          .value(),
+      )
+      // Loadem and loggem
+      .then((tasks) => _.flatten(tasks))
+      .each((task) => {
+        lando.tasks.push(require(task)(lando));
+        lando.log.debug('autoloaded task %s', path.basename(task, '.js'));
+      })
+      // Reset the task cache
+      .then(() => {
+        lando.cache.set('_.tasks.cache', JSON.stringify(lando.tasks), { persist: true });
+      })
+  );
 };
 
 /*
  * Helper to bootstrap engine
  */
-const bootstrapEngine = lando => {
+const bootstrapEngine = (lando) => {
   const Shell = require('./shell');
   lando.shell = new Shell(lando.log);
   lando.scanUrls = require('./scan')(lando.log);
@@ -112,7 +126,7 @@ const bootstrapEngine = lando => {
   );
   lando.utils = _.merge({}, require('./utils'), require('./config'));
   // Auto move and make executable any scripts
-  return lando.Promise.map(lando.config.plugins, plugin => {
+  return lando.Promise.map(lando.config.plugins, (plugin) => {
     if (fs.existsSync(plugin.scripts)) {
       const confDir = path.join(lando.config.userConfRoot, 'scripts');
       const dest = lando.utils.moveConfig(plugin.scripts, confDir);
@@ -125,20 +139,20 @@ const bootstrapEngine = lando => {
 /*
  * Helper to bootstrap app stuffs
  */
-const bootstrapApp = lando => {
+const bootstrapApp = (lando) => {
   const Factory = require('./factory');
-  const Yaml = require('./yaml');
+  const Yaml = require('../lib/yaml');
   lando.factory = new Factory();
   lando.yaml = new Yaml(lando.log);
   // Load in all our builders in the correct order
   const builders = _(['compose', 'types', 'services', 'recipes'])
-    .flatMap(type => _.map(lando.config.plugins, plugin => plugin[type]))
-    .filter(dir => fs.existsSync(dir))
-    .flatMap(dir => glob.sync(path.join(dir, '*', 'builder.js')))
-    .map(file => lando.factory.add(require(file)).name)
+    .flatMap((type) => _.map(lando.config.plugins, (plugin) => plugin[type]))
+    .filter((dir) => fs.existsSync(dir))
+    .flatMap((dir) => glob.sync(path.join(dir, '*', 'builder.js')))
+    .map((file) => lando.factory.add(require(file)).name)
     .value();
   // Log
-  _.forEach(builders, builder => lando.log.debug('autoloaded builder %s', builder));
+  _.forEach(builders, (builder) => lando.log.debug('autoloaded builder %s', builder));
 };
 
 /*
@@ -146,11 +160,16 @@ const bootstrapApp = lando => {
  */
 const bootstrapRouter = (level, lando) => {
   switch (level) {
-    case 'config': return bootstrapConfig(lando);
-    case 'tasks': return bootstrapTasks(lando);
-    case 'engine': return bootstrapEngine(lando);
-    case 'app': return bootstrapApp(lando);
-    default: return true;
+    case 'config':
+      return bootstrapConfig(lando);
+    case 'tasks':
+      return bootstrapTasks(lando);
+    case 'engine':
+      return bootstrapEngine(lando);
+    case 'app':
+      return bootstrapApp(lando);
+    default:
+      return true;
   }
 };
 
@@ -192,8 +211,7 @@ module.exports = class Lando {
     this.cache = bootstrap.setupCache(this.log, this.config);
     this.log = new Log(this.config);
     this.metrics = bootstrap.setupMetrics(this.log, this.config);
-    this.error = new ErrorHandler(this.log, this.metrics),
-    this.events = new AsyncEvents(this.log);
+    (this.error = new ErrorHandler(this.log, this.metrics)), (this.events = new AsyncEvents(this.log));
     this.updates = new UpdateManager();
     this.user = require('./user');
   }
@@ -243,13 +261,13 @@ module.exports = class Lando {
   bootstrap(level = 'app') {
     // Log that we've begun
     this.log.verbose('starting bootstrap at level %s...', level);
-    this.log.silly('it\'s not particularly silly, is it?');
+    this.log.silly("it's not particularly silly, is it?");
 
     // @TODO TEST THE BELOW BIG TIMEZ
     const bootstraps = _.slice(_.keys(BOOTSTRAP_LEVELS), 0, BOOTSTRAP_LEVELS[level]);
 
     // Loop through our bootstrap levels
-    return this.Promise.each(bootstraps, level => {
+    return this.Promise.each(bootstraps, (level) => {
       this.log.verbose('%s bootstrap beginning...', level);
       this._bootstrap = level;
 
@@ -301,66 +319,69 @@ module.exports = class Lando {
        *   // My codes
        * });
        */
-      return this.events.emit(`pre-bootstrap-${level}`, this)
+      return (
+        this.events
+          .emit(`pre-bootstrap-${level}`, this)
 
-      // Call the things that should happen at each level
-      .then(() => bootstrapRouter(level, this))
+          // Call the things that should happen at each level
+          .then(() => bootstrapRouter(level, this))
 
-      /**
-       * Event that runs after we bootstrap config
-       *
-       * @since 3.0.0
-       * @alias lando.events:post-bootstrap-config
-       * @event post_bootstrap_config
-       * @property {Lando} lando The Lando object
-       * @example
-       * lando.events.on('post-bootstrap-config', lando => {
-       *   // My codes
-       * });
-       */
-      /**
-       * Event that runs after we bootstrap tasks
-       *
-       * @since 3.0.0
-       * @alias lando.events:post-bootstrap-tasks
-       * @event post_bootstrap_tasks
-       * @property {Lando} lando The Lando object
-       * @example
-       * lando.events.on('post-bootstrap-tasks', lando => {
-       *   // My codes
-       * });
-       */
-      /**
-       * Event that runs after we bootstrap the engine
-       *
-       * @since 3.0.0
-       * @alias lando.events:post-bootstrap-engine
-       * @event post_bootstrap_engine
-       * @property {Lando} lando The Lando object
-       * @example
-       * lando.events.on('post-bootstrap-engine', lando => {
-       *   // My codes
-       * });
-       */
-      /**
-       * Event that runs after we bootstrap the app
-       *
-       * @since 3.0.0
-       * @alias lando.events:post-bootstrap-app
-       * @event post_bootstrap_app
-       * @property {Lando} lando The Lando object
-       * @example
-       * lando.events.on('post-bootstrap-app', lando => {
-       *   // My codes
-       * });
-       */
-      .then(() => this.events.emit(`post-bootstrap-${level}`, this))
-      // Log the doneness
-      .then(() => this.log.verbose('%s bootstrap completed.', level));
+          /**
+           * Event that runs after we bootstrap config
+           *
+           * @since 3.0.0
+           * @alias lando.events:post-bootstrap-config
+           * @event post_bootstrap_config
+           * @property {Lando} lando The Lando object
+           * @example
+           * lando.events.on('post-bootstrap-config', lando => {
+           *   // My codes
+           * });
+           */
+          /**
+           * Event that runs after we bootstrap tasks
+           *
+           * @since 3.0.0
+           * @alias lando.events:post-bootstrap-tasks
+           * @event post_bootstrap_tasks
+           * @property {Lando} lando The Lando object
+           * @example
+           * lando.events.on('post-bootstrap-tasks', lando => {
+           *   // My codes
+           * });
+           */
+          /**
+           * Event that runs after we bootstrap the engine
+           *
+           * @since 3.0.0
+           * @alias lando.events:post-bootstrap-engine
+           * @event post_bootstrap_engine
+           * @property {Lando} lando The Lando object
+           * @example
+           * lando.events.on('post-bootstrap-engine', lando => {
+           *   // My codes
+           * });
+           */
+          /**
+           * Event that runs after we bootstrap the app
+           *
+           * @since 3.0.0
+           * @alias lando.events:post-bootstrap-app
+           * @event post_bootstrap_app
+           * @property {Lando} lando The Lando object
+           * @example
+           * lando.events.on('post-bootstrap-app', lando => {
+           *   // My codes
+           * });
+           */
+          .then(() => this.events.emit(`post-bootstrap-${level}`, this))
+          // Log the doneness
+          .then(() => this.log.verbose('%s bootstrap completed.', level))
+      );
     })
-    .then(() => this.log.verbose('bootstrap completed.'))
-    .then(() => this.events.emit(`post-bootstrap`, this))
-    .then(() => this);
+      .then(() => this.log.verbose('bootstrap completed.'))
+      .then(() => this.events.emit(`post-bootstrap`, this))
+      .then(() => this);
   }
 
   /**
@@ -379,7 +400,7 @@ module.exports = class Lando {
   getApp(startFrom = process.cwd(), warn = true) {
     const merger = require('./config').merge;
     const utils = require('./bootstrap');
-    const Yaml = require('./yaml');
+    const Yaml = require('../lib/yaml');
     const yaml = new Yaml(this.log);
     // Grab lando files for this app
     const fileNames = _.flatten([this.config.preLandoFiles, [this.config.landoFile], this.config.postLandoFiles]);
@@ -393,11 +414,11 @@ module.exports = class Lando {
     }
 
     // Load the config and augment so we can get an App
-    const config = merger({}, ..._.map(landoFiles, file => yaml.load(file)));
+    const config = merger({}, ..._.map(landoFiles, (file) => yaml.load(file)));
     this.log.info('loading app %s from config files', config.name, landoFiles);
     this.log.debug('app %s has config', config.name, config);
     // Return us some app!
     const App = require('./app');
-    return new App(config.name, _.merge({}, config, {files: landoFiles}), this);
+    return new App(config.name, _.merge({}, config, { files: landoFiles }), this);
   }
 };
