@@ -1,3 +1,3 @@
 export default async ({ debug }) => {
-  debug('hello there!');
+  debug('hello there2!');
 };
