@@ -1,20 +1,26 @@
-'use strict';
+import Debug from 'debug';
 
 // normal debug module with some extra helper funcs
-module.exports = namespace => {
+function createDebug(namespace) {
   // get the normal debugger
-  const debug = require('debug')(namespace);
+  const debug = Debug(namespace);
 
   // this does the opposite of extend, moves the namespace back by size delimiter
-  debug.contract = size => module.exports(debug.namespace.split(':').slice(0, Math.abs(size) * -1).join(':'));
+  debug.contract = (size) =>
+    createDebug(
+      debug.namespace
+        .split(':')
+        .slice(0, Math.abs(size) * -1)
+        .join(':'),
+    );
 
   // keep the old extend
   debug._extend = debug.extend;
   // shallow wrapper just so debug.extend gives us contract and replace as well
-  debug.extend = more => module.exports([debug.namespace, more].join(':'));
+  debug.extend = (more) => createDebug([debug.namespace, more].join(':'));
 
   // replaces the string or component piece
-  // @NOTE: this assumes a first value of 1 and not 0
+  // @NOTE: this assumes a first value of  1 and not 0
   // @NOTE: but should we assume that?
   // @TODO: error handling?
   debug.replace = (replacee, replacer) => {
@@ -25,11 +31,11 @@ module.exports = namespace => {
     if (Number.isInteger(replacee)) {
       const parts = debug.namespace.split(':');
       parts[replacee - 1] = replacer;
-      return module.exports(parts.join(':'));
+      return createDebug(parts.join(':'));
 
-    // just do a simple string replacement
+      // just do a simple string replacement
     } else if (typeof replacee === 'string') {
-      return module.exports('debug')(debug.namespace.replace(replacee, replacer));
+      return createDebug('debug')(debug.namespace.replace(replacee, replacer));
     }
 
     // return
@@ -37,4 +43,6 @@ module.exports = namespace => {
   };
 
   return debug;
-};
+}
+
+export default createDebug;
