@@ -25,10 +25,7 @@ class Product {
     this.#_init();
   }
 
-  constructor(config, {
-    debug = Product.debug,
-    StorageBackend = require('../components/file-storage'),
-  } = {}) {
+  constructor(config, { debug = Product.debug, StorageBackend = require('../components/file-storage') } = {}) {
     // @TODO: throw error if not a config class?
     // @TODO: options for config?
     this.config = config || new require('./config')(); // eslint-disable-line new-cap
@@ -46,11 +43,11 @@ class Product {
     this.debug = debug;
 
     // add our defaults as a source
-    this.config.defaults('product-defaults', require('../config/product-defaults')({id: this.id, env: this.id}));
+    this.config.defaults('product-defaults', require('../workspace/product-defaults')({ id: this.id, env: this.id }));
 
     // get the id
     if (!this.config.get(`${this.config.managed}:system.instance`)) {
-      const data = {system: {instance: require('../utils/generate-id')()}};
+      const data = { system: { instance: require('../utils/generate-id')() } };
       this.config.save(data);
       this.config.set('system.instance', data.system.instance);
       this.debug('could not locate instance id, setting to %o', this.config.get('system.instance'));
@@ -60,7 +57,10 @@ class Product {
     // we do this because we need a way a to cache things before plugins/the registry are compiled,
     // because we cannot use something like async getComponent inside the constructor and because we do not
     // want a full-blow async init() method we need to call EVERY time we new Bootstrap()
-    this.#_cache = new StorageBackend(({debug: this.debug.extend('#cache'), dir: this.config.get('system.syscache-dir')}));
+    this.#_cache = new StorageBackend({
+      debug: this.debug.extend('#cache'),
+      dir: this.config.get('system.syscache-dir'),
+    });
 
     // if caching is disable and cache dir exists then flush
     if (!this.config.get('core.caching') && fs.existsSync(this.config.get('system.syscache-dir'))) {
@@ -72,9 +72,11 @@ class Product {
 
     // "private" but still "by convention" protected properties
     this._componentsCache = {};
-    this._componentAlises = Object.fromEntries(Object.entries(this.config.getUncoded('core'))
-      .filter(([type, component]) => typeof component === 'string')
-      .map(([type, component]) => ([`core.${type}`, `${type}.${component}`])));
+    this._componentAlises = Object.fromEntries(
+      Object.entries(this.config.getUncoded('core'))
+        .filter(([type, component]) => typeof component === 'string')
+        .map(([type, component]) => [`core.${type}`, `${type}.${component}`]),
+    );
   }
 
   // @TODO: the point of this is to have a high level way to "fetch" a certain kind of plugin eg global and
@@ -118,7 +120,7 @@ class Product {
     // @TODO: should we allow for a separate aliases section in the config?
     return require('../utils/get-component')(
       component,
-      this.config.newConfig({data: this.getRegistry(), id: `${this.id}-component-registry`}),
+      this.config.newConfig({ data: this.getRegistry(), id: `${this.id}-component-registry` }),
       {
         aliases: this._componentAlises,
         cache: this._componentsCache,
@@ -130,24 +132,21 @@ class Product {
 
   // helper to get a component (and config?) from the registry
   async getComponentInstance(component, constructor = {}, init = true) {
-    return require('../utils/get-component-instance')(
-      component,
-      constructor,
-      {
-        aliases: this._componentAlises,
-        cache: this._componentsCache,
-        config: this.config.get(),
-        debug: this.debug.extend('get-component-instance'),
-        registry: this.config.newConfig({data: this.getRegistry(), id: `${this.id}-component-registry`}),
-      },
-    );
+    return require('../utils/get-component-instance')(component, constructor, {
+      aliases: this._componentAlises,
+      cache: this._componentsCache,
+      config: this.config.get(),
+      debug: this.debug.extend('get-component-instance'),
+      registry: this.config.newConfig({ data: this.getRegistry(), id: `${this.id}-component-registry` }),
+    });
   }
 
   getHooks() {
-    const hooks = this.config.merge(this.hooks,
+    const hooks = this.config.merge(
+      this.hooks,
       [
-        this.manifest.getUncoded('hooks.product', {ams: 'aoa'}),
-        this.manifest.getUncoded(`hooks.${this.id}`, {ams: 'aoa'}),
+        this.manifest.getUncoded('hooks.product', { ams: 'aoa' }),
+        this.manifest.getUncoded(`hooks.${this.id}`, { ams: 'aoa' }),
       ],
       ['aoa'],
     );
@@ -165,7 +164,7 @@ class Product {
     this.debug('building %o manifest...', this.id);
 
     // build the manifest from our plugins
-    const manifest = this.config.newConfig({id: 'manifest'});
+    const manifest = this.config.newConfig({ id: 'manifest' });
     // loop through plugins in reverse order and add them to the manifest config
     for (const [name, plugin] of Object.entries(this.plugins).reverse()) {
       // clone the manifest
@@ -176,7 +175,7 @@ class Product {
       delete store.description;
       delete store.enabled;
       delete store.hidden;
-      manifest.add(name, {type: 'literal', store});
+      manifest.add(name, { type: 'literal', store });
     }
 
     // set
@@ -202,7 +201,7 @@ class Product {
     this.debug('running %o plugin discovery...', this.id);
 
     // do the discovery
-    const {disabled, enabled, invalids} = require('../utils/get-plugins')(
+    const { disabled, enabled, invalids } = require('../utils/get-plugins')(
       this.config.get('plugin.dirs'),
       {
         config: this.config.get(),
@@ -226,7 +225,7 @@ class Product {
   }
 
   getRegistry() {
-    return this.manifest.get('registry', {decode: false, encode: false});
+    return this.manifest.get('registry', { decode: false, encode: false });
   }
 
   reinit() {
@@ -254,7 +253,7 @@ class Product {
   }
 
   async runHook(event, data) {
-    return require('../utils/run-hook')(event, data, this.hooks, {[this.id]: this, product: this}, this.debug);
+    return require('../utils/run-hook')(event, data, this.hooks, { [this.id]: this, product: this }, this.debug);
   }
 }
 
